Support cgroup v2 Fixes #27
